Pathfinder re-route implementation
At any point, pressing "r" causes the pathfinder to calculate a new route considering the demolished wall by the wanderer and put it in effect.

There is a problem in the re-calculated paths as they no longer abide the "hall" case, with possible "squares"/"openings". Recorded this issue as TODO in eMaze.py